move executor builder to submodule
